fix: resolve critical loading and functionality issues in refactored camera component
This commit addresses multiple critical issues discovered during testing of the refactored camera component, ensuring functional parity with the original implementation while maintaining the improved modular architecture.

## Key Fixes:

### 🚨 Critical Loading Issue
- Add missing loading overlay (wd-overlay) that provides visual feedback during interview initialization
- Fix ReferenceError: domObserver is not defined by properly importing useDOMObserver
- Implement robust toast.loading fallback mechanism with error handling

### 🔄 Interview Flow Corrections
- Restore interview data re-fetch on startInterview() to match original behavior
- Fix API response handling to correctly access backend data structure
- Correct initialization sequence: Interview Info → Camera → Hide Video Mask

### 🎥 Upload & Media Handling
- Simplify upload error handling to match original behavior (show toast, stay on question)
- Remove complex retry/skip dialog that could cause data loss
- Restore TTS error feedback for missing audio URLs
- Remove duplicate DOM observers to prevent conflicts

### 🛠️ Technical Improvements
- Fix backend API response structure handling (interview_id from params, not response)
- Ensure proper cleanup of all composables and observers
- Maintain modular architecture while restoring critical functionality

## Backend API Compatibility:
- Updated frontend to correctly handle backend response structure
- Backend returns {code: 200, data: {position, questions}} format
- interview_id comes from URL parameter, not API response

## Testing Status:
- ✅ Build completes successfully with no TypeScript errors
- ✅ Loading states properly managed with visual feedback
- ✅ Interview initialization and camera setup working correctly
- ✅ Upload error handling matches original safe behavior
- ✅ Exit/terminate flows work as expected

Fixes issues identified in colleague assessment and ensures the refactored component maintains exact functional parity with the production-ready original implementation.
