+++ src/pages/camera/FIX_INTERVIEW_FLOW.md
@@ -1,0 +1,33 @@
+# 面试流程修复方案
+
+## 已识别的问题
+
+1. **API路径错误**：已修复为 `/files/post-policy`
+2. **COS上传参数缺失**：已修复，添加了完整的COS上传逻辑
+3. **音频播放错误处理**：已有超时保护机制
+4. **[completeInterview] Error: TypeError**：需要修复面试完成逻辑
+
+## 主要问题分析
+
+### 1. 上传失败的根本原因
+- API路径已修正为 `/files/post-policy?ext=webm`
+- 上传服务已重写为完整的COS直传模式
+- 包含了XMLHttpRequest备用方案
+
+### 2. 面试完成错误
+错误日志显示 `[completeInterview] Error: TypeError`，可能原因：
+- 访问了未定义的属性
+- 异步操作的时序问题
+- Store状态未正确初始化
+
+## 需要进一步检查的地方
+
+1. **InterviewControls组件**：确保正确判断最后一题
+2. **面试完成时的重定向处理**：检查跳转逻辑
+3. **视频上传成功后的回调**：确保正确保存到store
+
+## 修复建议
+
+1. 在 `index.vue` 中添加更多的错误处理
+2. 确保所有异步操作都有适当的错误边界
+3. 检查 store 中的数据是否完整初始化
