+++ src/pages/camera/components/InterviewControls.vue
@@ -2,21 +2,15 @@
   <view class="interview-controls">
     <!-- 未开始时的按钮 -->
     <template v-if="!isStarted">
-      <view class="control-btn exit-btn" @click="handleExit">
-        退出面试
-      </view>
-      <view class="control-btn start-btn" @click="handleStart">
-        开始面试
-      </view>
+      <view class="control-btn exit-btn" @click="handleExit">退出面试</view>
+      <view class="control-btn start-btn" @click="handleStart">开始面试</view>
     </template>
-    
+
     <!-- 进行中的按钮 -->
     <template v-else>
-      <view class="control-btn exit-btn" @click="handleTerminate">
-        终止面试
-      </view>
-      <view 
-        class="control-btn next-btn" 
+      <view class="control-btn exit-btn" @click="handleTerminate">终止面试</view>
+      <view
+        class="control-btn next-btn"
         :class="{ 'btn-disabled': isDisabled }"
         @click="handleNext"
       >
@@ -38,7 +32,7 @@ interface Props {
 
 const props = withDefaults(defineProps<Props>(), {
   isLastQuestion: false,
-  isLoading: false
+  isLoading: false,
 })
 
 const emit = defineEmits<{
@@ -70,17 +64,7 @@ const handleExit = () => {
 
 const handleTerminate = () => {
   if (!props.isLoading) {
-    uni.showModal({
-      title: '提示',
-      content: '确定要终止面试吗？终止后将无法继续答题。',
-      confirmText: '终止',
-      confirmColor: '#ff4444',
-      success: (res) => {
-        if (res.confirm) {
-          emit('terminate')
-        }
-      }
-    })
+    emit('terminate')
   }
 }
 
@@ -94,17 +78,17 @@ const handleNext = () => {
 <style scoped>
 .interview-controls {
   position: fixed;
-  bottom: 80px;
+  bottom: 80px; /* 原版使用 bottom-20 = 20 * 4px = 80px */
   left: 0;
   right: 0;
   display: flex;
   justify-content: center;
-  gap: 20px;
+  padding: 0 10px;
+  gap: 8px;
   z-index: 10;
 }
 
 .control-btn {
-  width: 35%;
   padding: 10px 0;
   border-radius: 8px;
   text-align: center;
@@ -118,18 +102,21 @@ const handleNext = () => {
 }
 
 .exit-btn {
+  width: 35%;
   background-color: #fff;
   color: #3b82f6;
   border: 1px solid #3b82f6;
 }
 
 .start-btn {
-  background-color: #fd7e11;
+  width: 60%;
+  background-color: #3b82f6;
   color: #fff;
 }
 
 .next-btn {
-  background-color: #fd7e11;
+  width: 60%;
+  background-color: #10b981;
   color: #fff;
 }
