+++ src/pages/interviews/process-simulation.vue
@@ -132,6 +132,7 @@ import { onPageScroll as uniPageScroll } from '@dcloudio/uni-app'
 import { useToast } from 'wot-design-uni'
 import { navigateBack } from '@/utils/platformUtils'
  import { handleToken } from "@/utils/useAuth"
+
 const toast = useToast()
 const showSheet = ref(false)
 const items = ref([])
