+++ src/pages/interviews/record-simulate.vue
@@ -144,7 +144,7 @@ import dw from '../../static/app/icons/icon_dw.png'
 import dh from '../../static/app/icons/icon_dh.png'
 import { useQueue, useToast, useMessage } from 'wot-design-uni'
 import { navigateBack } from '@/utils/platformUtils'
-import { handleToken } from "@/utils/useAuth"
+import { handleToken } from '@/utils/useAuth'
 import { ref, watch, onMounted } from 'vue'
 const toast = useToast()
 
@@ -166,21 +166,25 @@ onMounted(() => {
   console.log('record-simulate.vue - interviewList初始值:', interviewList.value)
   getPostionInfo()
   my_test_interviews()
-  
+
   // 添加watch来监控interviewList的变化
-  watch(interviewList, (newVal, oldVal) => {
-    console.log('interviewList发生变化:')
-    console.log('- 旧值长度:', oldVal?.length || 0)
-    console.log('- 新值长度:', newVal?.length || 0)
-    console.log('- 新值内容:', JSON.stringify(newVal, null, 2))
-  }, { deep: true })
+  watch(
+    interviewList,
+    (newVal, oldVal) => {
+      console.log('interviewList发生变化:')
+      console.log('- 旧值长度:', oldVal?.length || 0)
+      console.log('- 新值长度:', newVal?.length || 0)
+      console.log('- 新值内容:', JSON.stringify(newVal, null, 2))
+    },
+    { deep: true },
+  )
 })
 
 const my_test_interviews = async (keyword = '') => {
   // 在函数开始处打印token值
   const token = uni.getStorageSync('token')
   console.log('my_test_interviews - 当前token值:', token)
-  
+
   if (keyword.trim() !== '') {
     console.log('search')
   }
@@ -189,10 +193,10 @@ const my_test_interviews = async (keyword = '') => {
   const trimmedKeyword = keyword.trim()
   const queryParams = trimmedKeyword ? `?keyword=${encodeURIComponent(trimmedKeyword)}` : ''
   const url = `${baseUrl}/interviews/my_test_interviews${queryParams}`
-  
+
   try {
     const response = await uni.request({
-      url: url,
+      url,
       method: 'GET',
       header: { Authorization: `Bearer ${uni.getStorageSync('token')}` },
     })
@@ -201,7 +205,10 @@ const my_test_interviews = async (keyword = '') => {
     if (response.statusCode === 200) {
       interviewList.value = response.data.data
       console.log('my_test_interviews - interviewList已更新，长度:', interviewList.value.length)
-      console.log('my_test_interviews - interviewList内容:', JSON.stringify(interviewList.value, null, 2))
+      console.log(
+        'my_test_interviews - interviewList内容:',
+        JSON.stringify(interviewList.value, null, 2),
+      )
       // 检查每个item的interviews_id
       interviewList.value.forEach((item, index) => {
         console.log(`面试记录[${index}] - interviews_id:`, item.interviews_id, '完整数据:', item)
@@ -210,11 +217,11 @@ const my_test_interviews = async (keyword = '') => {
   } catch (error) {
     // 添加更详细的错误信息打印
     console.error('my_test_interviews请求失败 - 详细错误信息:', {
-      error: error,
+      error,
       errorMessage: error?.message || '未知错误',
       errorStack: error?.stack || '无堆栈信息',
-      url: url,
-      token: uni.getStorageSync('token')
+      url,
+      token: uni.getStorageSync('token'),
     })
     toast.error('面试结果正在生成中，请稍后再试')
   } finally {
@@ -225,50 +232,52 @@ const my_test_interviews = async (keyword = '') => {
 const getPostionInfo = async () => {
   // 记录函数开始执行
   console.log('=== getPostionInfo 开始执行 ===')
-  
+
   // 构建完整URL并记录
   const url = baseUrl + `/jobseekers/by-user/`
   const token = uni.getStorageSync('token')
-  
+
   console.log('getPostionInfo - 请求详情:', {
-    url: url,
-    baseUrl: baseUrl,
+    url,
+    baseUrl,
     token: token ? `Bearer ${token.substring(0, 10)}...` : '无token',
-    method: 'GET'
+    method: 'GET',
   })
-  
+
   try {
     const response = await uni.request({
-      url: url,
+      url,
       method: 'GET',
       header: { Authorization: `Bearer ${token}` },
     })
-    
+
     console.log('getPostionInfo - 响应详情:', {
       statusCode: response.statusCode,
       dataType: typeof response.data,
       dataIsArray: Array.isArray(response.data),
       dataLength: Array.isArray(response.data) ? response.data.length : 'N/A',
-      sampleData: response.data ? JSON.stringify(response.data).substring(0, 200) + '...' : '无数据'
+      sampleData: response.data
+        ? JSON.stringify(response.data).substring(0, 200) + '...'
+        : '无数据',
     })
-    
+
     if (response.statusCode === 200) {
       // 清空items数组以避免重复
       items.value = []
-      
+
       // 检查响应数据格式
       if (!response.data) {
         console.warn('getPostionInfo - 响应数据为空')
         return
       }
-      
+
       if (!Array.isArray(response.data)) {
         console.warn('getPostionInfo - 响应数据不是数组，实际类型:', typeof response.data)
         return
       }
-      
+
       console.log(`getPostionInfo - 开始处理 ${response.data.length} 个职位`)
-      
+
       response.data.forEach((element, index) => {
         console.log(`处理职位[${index}]:`, {
           position_name: element.position_name,
@@ -276,9 +285,9 @@ const getPostionInfo = async () => {
           id: element.id,
           expected_city: element.expected_city,
           salary_min: element.expected_salary_min,
-          salary_max: element.expected_salary_max
+          salary_max: element.expected_salary_max,
         })
-        
+
         let salaryStr = ''
         if (element.expected_salary_min === '待议' && element.expected_salary_max === '待议') {
           salaryStr = ''
@@ -290,7 +299,7 @@ const getPostionInfo = async () => {
         } else {
           salaryStr = element.expected_salary_min + '-' + element.expected_salary_max
         }
-        
+
         const itemData = {
           title: element.position_name,
           description: element.position_name,
@@ -301,22 +310,21 @@ const getPostionInfo = async () => {
           expected_city: element.expected_city,
           id: element.id,
         }
-        
+
         items.value.push(itemData)
         console.log(`职位[${index}]处理完成，当前items长度:`, items.value.length)
       })
-      
+
       console.log('getPostionInfo - 处理完成，最终items:', {
         length: items.value.length,
-        items: items.value
+        items: items.value,
       })
-      
     } else {
       console.error('getPostionInfo - 请求失败:', {
         statusCode: response.statusCode,
         statusText: response.statusText || '无状态文本',
         data: response.data,
-        headers: response.headers || '无响应头'
+        headers: response.headers || '无响应头',
       })
     }
   } catch (error) {
@@ -326,11 +334,11 @@ const getPostionInfo = async () => {
       errorCode: error?.code || '无错误代码',
       errorDetail: error?.detail || '无详细信息',
       errorStack: error?.stack || '无堆栈信息',
-      url: url,
+      url,
       token: token ? '有token' : '无token',
-      fullError: error
+      fullError: error,
     })
-    
+
     // 根据不同错误类型给出更具体的提示
     if (error?.code === 'NETWORK_ERROR' || error?.errMsg?.includes('network')) {
       toast.error('网络连接失败，请检查网络设置')
@@ -340,7 +348,7 @@ const getPostionInfo = async () => {
       toast.error('获取职位信息失败，请稍后重试')
     }
   }
-  
+
   console.log('=== getPostionInfo 执行结束 ===')
 }
 onLoad((options) => {
@@ -392,7 +400,7 @@ const openInfo = (id) => {
   console.log('openInfo - id是否为undefined:', id === undefined)
   console.log('openInfo - id是否为null:', id === null)
   console.log('openInfo - id是否为0:', id === 0)
-  
+
   if (!id || id === undefined || id === null) {
     console.log('openInfo - id无效，显示错误提示')
     toast.error('面试结果正在生成中，请稍后再试')
@@ -425,9 +433,9 @@ const submitTestInerview = async () => {
         },
         fail: (error) => {
           console.error('submitTestInerview - 失败详情:', {
-            error: error,
+            error,
             token: uni.getStorageSync('token'),
-            selectedItemId: selectedItem.id
+            selectedItemId: selectedItem.id,
           })
         },
         complete: () => {
