+++ src/typings/global.d.ts
@@ -9,4 +9,4 @@ interface Window {
       }
     }
   }
-} 
+}
