--- test_audio_error_handling.md
@@ -1,46 +1,0 @@
-# 音频错误处理测试用例
-
-## 测试场景
-
-### 1. 音频URL为空或null的情况
-- **预期行为**: 应该跳过音频播放，给用户3秒时间阅读题目，然后继续面试流程
-- **日志输出**: "当前问题没有音频URL，跳过播放，直接开始面试"
-
-### 2. 音频URL格式无效
-- **预期行为**: 捕获错误并继续面试流程
-- **日志输出**: "无效的音频URL格式:" 和 "音频播放设置失败:"
-
-### 3. 音频加载超时（10秒）
-- **预期行为**: 10秒后自动停止等待，显示错误提示，继续面试
-- **用户提示**: "音频加载超时，请阅读题目"
-- **日志输出**: "音频播放超时（10秒），跳过音频继续面试"
-
-### 4. 音频播放错误
-- **预期行为**: 根据错误类型显示相应日志，给用户提示，然后继续面试
-- **用户提示**: "音频播放失败，请阅读题目"
-- **错误类型**:
-  - 10001: 系统错误
-  - 10002: 网络错误
-  - 10003: 文件错误
-  - 10004: 格式错误
-  - -1: 未知错误
-
-### 5. 音频正常播放
-- **预期行为**: 音频播放结束后自动开始面试
-- **日志输出**: "音频播放结束，开始面试"
-
-## 关键改进点
-
-1. **容错性**: 任何音频相关的错误都不会阻止面试继续进行
-2. **超时保护**: 10秒超时机制防止无限等待
-3. **清理机制**: 切换题目和组件卸载时正确清理音频资源和计时器
-4. **用户体验**: 适当的错误提示让用户知道发生了什么，但不影响面试进行
-5. **日志记录**: 详细的日志帮助调试问题
-
-## 测试方法
-
-1. 模拟TTS服务失败（audio_url为null）
-2. 提供无效的音频URL格式
-3. 提供一个无法访问的音频URL
-4. 在音频播放过程中快速切换题目
-5. 在音频播放过程中退出页面
